Fix type definition to reflect CommonJS module export
This fixes an issue around using this module's type definition in combination with JavaScript code that is type checked via the typescript compiler.

The following code, which works just fine when executed, currently does not type check correctly:

```js
// @ts-check
const AbortController = require('node-abort-controller');
new AbortController();
```

This is the error message that is generated by `tsc` on line 3:

```
This expression is not constructable.
  Type 'typeof import("/workspaces/tedious/node_modules/node-abort-controller/index")' has no construct signatures.
```

The reason for this is that the type definition uses `export default`, which is different from the `module.exports = AbortController` that actually happens inside `node-abort-controller`. Lucky for us, TypeScript actually has a special syntax to correctly model definitions for CommonJS libraries, and that is `exports =`. By changing from `export default ...`, we can ensure that this module can be used correctly from TypeScript code that uses `import AbortController from 'node-abort-controller';` as well as JavaScript code that uses `const AbortController = require('node-abort-controller')`.

Loading the `AbortSignal` class is untouched and continues working as, via `import { AbortSignal } from 'node-abort-controller';` as well as `const AbortSignal = require('node-abort-controller').AbortSignal;`.